Move user related functions to account app
