+++ calliope_app/account/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class AccountConfig(AppConfig):
+    name = 'account'
