+++ calliope_app/client/templates/registration/registration_form.html
@@ -1,0 +1,10 @@
+{% load i18n %}
+{% load crispy_forms_tags %}
+
+<!--Register-->
+<form method="POST" id="registration_form" class="hide" data-action="{% url 'user_registration' %}">
+    {% csrf_token %}
+    {{ form | crispy }}
+    <button class="btn btn-sm btn-success" type="submit" name=""
+        value="{% trans 'Submit Registration' %}">Register</button>
+</form>
