+++ calliope_app/client/urls.py
@@ -93,15 +93,9 @@ urlpatterns = [
     path('share/',
          engage_views.share_view,
          name='settings'),
-    path('settings/password/',
-         engage_views.password_view,
-         name='password'),
     path('admin/login/',
          engage_views.admin_login_view,
          name="admin"),
-    path('login/',
-         engage_views.user_login,
-         name='login'),
     path('',
          engage_views.home_view,
          name='home')
